==> prep/01_crops_prep/step1_GAEZ_FAO_wrangling.Rmd <==
---
title: "GAEZ_wrangling"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Summary

Wrangle GAEZ crop category dataset.

# Data sources

**Reference**: Grogan, D., Frolking, S., Wisser, D. et al. Global gridded crop harvested area, production, yield, and monthly physical area data circa 2015. Sci Data 9, 15 (2022). https://doi.org/10.1038/s41597-021-01115-2

**Downloaded**: 07/14/2023

**Description**: "Here we provide an update to global gridded annual and monthly crop datasets. This new dataset uses the crop categories established by the Global Agro-Ecological Zones (GAEZ) Version 3 model, which is based on the Food and Agricultural Organization of the United Nations (FAO) crop production data. We used publicly available data from the FAOSTAT database as well as GAEZ Version 4 global gridded dataset to generate circa 2015 annual crop harvested area, production, and yields by crop production system (irrigated and rainfed) for 26 crops and crop categories globally at 5-minute resolution. We additionally used available data on crop rotations, cropping intensity, and planting and harvest dates to generate monthly gridded cropland data for physical areas for the 26 crops by production system. These data are in standard georeferenced gridded format, and can be used by any global hydrology, land surface, or other earth system model that requires gridded annual or monthly crop data inputs."

**Download Link**: https://gaez.fao.org/pages/data-access-download


**Reference**: FAOSTAT (2023) ‘Crops and livestock products’. Available at: https://www.fao.org/faostat/en/#data/QCL (Accessed: 14 March 2023).

**Downloaded**: 03/14/2023

**Description**: FAOSTAT crop production statistics 


# Setup

```{r setup, include = FALSE}
# getting packages we want
library(tidyverse)
library(here)
library(raster)
library(sf)
library(vroom)
library(janitor)

# Raster templates
source(here("src/spatial.R"))
source(here("src/fxns.R"))
source(here("src/directories.R"))

``` 


# Methods

```{r}

# read in GAEZ crop info data 

GAEZ_crops <- read.csv(here("prep/01_crops_prep/data/GAEZ_crop_info.csv")) %>%
  dplyr::select(1:3) %>%
  rename(item_code = FAOCODE) %>%
  filter(GAEZ_category != "Foddercrops")

GAEZ_crops$item_code <- gsub(" ", "", GAEZ_crops$item_code, fixed = TRUE)
GAEZ_crops$FAONAMES <- gsub(" ", "", GAEZ_crops$FAONAMES, fixed = TRUE)

```


## Splitting data 

```{r}
GAEZ_names <- GAEZ_crops %>%
   separate_rows(FAONAMES, item_code, sep = ",") %>%
  mutate(item_code = as.numeric(item_code))
```

## Check for duplicates

```{r}
GAEZ_names$item_code[duplicated(GAEZ_names$item_code)] #none.. good! 

```

## Save

```{r}
write_csv(GAEZ_names, here("prep/01_crops_prep/data/GAEZ_names.csv"))

```

## Make FAO crop dataset, with values for 2015 (GAEZ year) and 2021 (latest FAO reporting year)

```{r}
fao_crop_prod_raw <- readRDS(here("data/tidy_data/production-data/crops_production_tidy.rds"))

fao_crop_prod_2015_2021 <- fao_crop_prod_raw %>%
  dplyr::filter(element_code == 5510,
                year_code %in% c(2014, 2015, 2016, 2019, 2020, 2021)) %>%
  dplyr::select(sector, area_code, area, element_code, element, item_code, item, year_code, year, value)

write.csv(fao_crop_prod_2015_2021, here("prep/01_crops_prep/data/FAOSTAT_crop_production_2015_2021.csv"), row.names = FALSE)
```

## Check what crops are in each category.

```{r}
dsdf <- 
  vroom::vroom(here("prep/01_crops_prep/data/FAOSTAT_crop_production_2015_2021.csv")) %>% 
  clean_names() %>% 
  dplyr::select(item, item_code) %>% 
  unique() %>% 
  left_join(GAEZ_names) 

dsfdsf <- vroom::vroom(here("prep/01_crops_prep/data/FAO_crop_definitions.csv")) %>% 
  clean_names() %>% 
  dplyr::select(item_code, item, description) %>% 
  right_join(GAEZ_names)
```
